requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.0
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
twilio==7.17.0
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
python-multipart==0.0.6
pydantic[email]==1.10.7
aiofiles=23.1.0
